+++ CHANGES.md
@@ -49,6 +49,9 @@ may be available in the relevant data file.
 * `org.apache.sshd.agent.unix.AgentClient` constructor expects a non-*null* `FactoryManager` instance which
 it then exposes via its `getFactoryManager`.
 
+* `SftpEventListener#removing/removed` callbacks accept an `isDirectory` flag indicating the type of `Path` being
+removed - file or directory.
+
 ## Minor code helpers
 
 * The `Session` object provides a `isServerSession` method that can be used to distinguish between
@@ -145,3 +148,5 @@ and therefore closing all currently tracked file/directory handles.
 * [SSHD-921](https://issues.apache.org/jira/browse/SSHD-921) - Do not send session disconnect message due to timeout expiration if already done so.
 
 * [SSHD-923](https://issues.apache.org/jira/browse/SSHD-923) - Added agent close detection mechanisms to avoid infinite waits on incoming messages.
+
+* [SSHD-929](https://issues.apache.org/jira/browse/SSHD-929) - Provide file/directory flag indicator to SFTP event listener callback for removal.